requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2
openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
asyncio-mqtt==0.16.1
aiofiles==23.2.1
tqdm==4.67.1 